fix record header byte offset problem

+++ sql-insight-engine/sql-insight-engine-innodb/src/main/kotlin/tech/insight/engine/innodb/page/SystemUserRecord.kt
@@ -184,7 +184,7 @@ class Infimum private constructor(private val belongToPage: InnoDbPage) : System
     }
 
     override fun nextRecordOffset(): Int {
-        return recordHeader.nextRecordOffset
+        return recordHeader.nextRecordOffset.toInt()
     }
 
     override fun absoluteOffset(): Int {
